feat: move/update/fix tests
