+++ .github/workflows/e2e-tests.yml
@@ -1,4 +1,4 @@
-name: Node.js CI
+name: E2E Tests
 permissions:
   contents: write
   packages: write
@@ -15,13 +15,14 @@ on:
         required: false
         default: false
 jobs:
-  test:
+  e2e-tests:
     if: github.actor != 'renovate[bot]'
     timeout-minutes: 60
     env:
       CI: true
       ESLINT_USE_FLAT_CONFIG: true
       SUPABASE_URL: http://localhost:54321
+      VITE_PORT: 5173
       VITE_BASE_URL: http://localhost:5173
       VITE_RECOVERY_AUTH_REDIRECT_URL: http://localhost:5173/profile/reset-password
       VITE_SUPABASE_URL: http://localhost:54321
