+++ .github/workflows/linting-and-unit-tests.yml
@@ -8,7 +8,7 @@ on:
   pull_request:
     branches: [main, staging]
 jobs:
-  test:
+  lint-and-unit-tests:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code ⬇️
