+++ tests/e2e/constants.ts
@@ -1,7 +1,6 @@
 import { createClient } from "@supabase/supabase-js";
 import "dotenv/config";
 
-export const baseUrl = process.env.VITE_BASE_URL;
 export const supabaseApiUrl = process.env.VITE_SUPABASE_URL ?? "";
 export const supabaseAnonKey = process.env.VITE_SUPABASE_ANON_KEY ?? "";
 export const inbucketUrl = process.env.TEST_SUPABASE_INBUCKET_URL ?? "";
