+++ tests/e2e/playwright.config.ts
@@ -1,17 +1,13 @@
 import { defineConfig, devices } from "@playwright/test";
 
-/**
- * Read environment variables from file.
- * https://github.com/motdotla/dotenv
- */
-// require('dotenv').config();
+const port = process.env.VITE_PORT ? parseInt(process.env.VITE_PORT) : 5173;
 
 /**
  * See https://playwright.dev/docs/test-configuration.
  */
 // eslint-disable-next-line @technologiestiftung/no-default-export
 export default defineConfig({
-	testDir: "./src/tests/e2e",
+	testDir: "./",
 	/* Run tests in files in parallel */
 	fullyParallel: true,
 	/* Fail the build on CI if you accidentally left test.only in the source code. */
@@ -21,16 +17,19 @@ export default defineConfig({
 	/* Opt out of parallel tests on CI. */
 	workers: 1,
 	/* Reporter to use. See https://playwright.dev/docs/test-reporters */
-	reporter: "list",
+	reporter: [
+		["list"],
+		["html", { open: "never", outputFolder: "./test-results" }],
+	],
 	/* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
 	use: {
 		/* Base URL to use in actions like `await page.goto('/')`. */
-		// baseURL: 'http://127.0.0.1:3000',
+		baseURL: process.env.VITE_BASE_URL,
 
 		/* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-		trace: "retain-on-failure",
+		trace: "on-first-retry",
 	},
-	timeout: 60000,
+	timeout: 30000,
 
 	/* Configure projects for major browsers */
 	projects: [
@@ -72,8 +71,16 @@ export default defineConfig({
 
 	/* Run your local dev server before starting the tests */
 	webServer: {
-		command: "npm run dev",
+		command: getCommand(),
 		url: process.env.VITE_BASE_URL,
 		reuseExistingServer: !process.env.CI,
 	},
 });
+
+function getCommand() {
+	if (process.env.CI) {
+		return `npm run preview -- --port ${port}`;
+	}
+
+	return `VITE_PW_TEST=true npm run build && npm run preview -- --port ${port}`;
+}
