+++ tests/unit/baby-tree-water-needs-special-district.test.ts
@@ -1,23 +1,23 @@
 import { expect, test } from "vitest";
-import { useTreeAgeClassification } from "../../components/tree-detail/hooks/use-tree-age-classification";
-import { useTreeWaterNeedsData } from "../../components/tree-detail/hooks/use-tree-water-needs-data";
+import { useTreeAgeClassification } from "../../src/components/tree-detail/hooks/use-tree-age-classification";
+import { useTreeWaterNeedsData } from "../../src/components/tree-detail/hooks/use-tree-water-needs-data";
 import {
 	TreeAgeClassification,
 	TreeCoreData,
-} from "../../components/tree-detail/tree-types";
+} from "../../src/components/tree-detail/tree-types";
 
-test("should calculate correct water needs for senior tree in special district", () => {
+test("should calculate correct water needs for baby tree in special district", () => {
 	const treeData: TreeCoreData = {
-		id: "00008100:0014f6e0",
-		lat: "13.40601",
-		lng: "52.53168",
-		art_dtsch: "Saulen-Hainbuche 'Frans Fontaine'",
-		art_bot: "Carpinus betulus 'Frans Fontaine'",
+		id: "00008100:00150353",
+		lat: "13.39502",
+		lng: "52.53667",
+		art_dtsch: "Pyramiden-Hainbuche",
+		art_bot: "Carpinus betulus 'Fastigiata'",
 		gattung_deutsch: "HAINBUCHE",
 		gattung: "CARPINUS",
-		pflanzjahr: 2011,
-		standalter: "12",
-		baumhoehe: "16",
+		pflanzjahr: 2012,
+		standalter: "11",
+		baumhoehe: "12",
 		bezirk: "Mitte",
 		eigentuemer: "Land Berlin",
 		radolan_sum: 281,
@@ -53,7 +53,7 @@ test("should calculate correct water needs for senior tree in special district",
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 4, 5, 10,
 		],
 		geom: {
-			coordinates: [13.40601, 52.53168],
+			coordinates: [13.39502, 52.53667],
 		},
 		caretaker: null,
 	};
@@ -62,8 +62,8 @@ test("should calculate correct water needs for senior tree in special district",
 		treeData,
 		new Date("2024-01-01T00:00+00:00"),
 	);
-	expect(treeAge).toBe(13);
-	expect(treeAgeClassification).toBe(TreeAgeClassification.SENIOR);
+	expect(treeAge).toBe(12);
+	expect(treeAgeClassification).toBe(TreeAgeClassification.BABY);
 
 	const {
 		rainSum,
@@ -78,12 +78,10 @@ test("should calculate correct water needs for senior tree in special district",
 
 	expect(rainSum).toBe(15.3);
 	expect(wateringSum).toBe(0);
-	expect(referenceWaterAmount).toBe(300);
-	expect(rainPercentage).toBe(0.051000000000000004);
+	expect(referenceWaterAmount).toBe(100);
+	expect(rainPercentage).toBe(0.153);
 	expect(wateringPercentage).toBe(0);
-	expect(otherWateringPercentage).toBe(0.949);
+	expect(otherWateringPercentage).toBe(0.847);
 	expect(shouldBeWatered).toBe(false);
-	expect(waterParts.map((p) => p.progress)).toEqual([
-		0.051000000000000004, 0, 0.949,
-	]);
+	expect(waterParts.map((p) => p.progress)).toEqual([0.153, 0, 0.847]);
 });
