+++ tests/unit/junior-tree-water-needs.test.ts
@@ -1,74 +1,74 @@
 import { expect, test } from "vitest";
-import { useTreeAgeClassification } from "../../components/tree-detail/hooks/use-tree-age-classification";
-import { useTreeWaterNeedsData } from "../../components/tree-detail/hooks/use-tree-water-needs-data";
+import { useTreeAgeClassification } from "../../src/components/tree-detail/hooks/use-tree-age-classification";
+import { useTreeWaterNeedsData } from "../../src/components/tree-detail/hooks/use-tree-water-needs-data";
 import {
 	TreeAgeClassification,
 	TreeCoreData,
 	TreeWateringData,
-} from "../../components/tree-detail/tree-types";
+} from "../../src/components/tree-detail/tree-types";
 
-test("should calculate correct water needs for baby tree", () => {
+test("should calculate correct water needs for junior tree", () => {
 	const waterings: TreeWateringData[] = [
 		{
 			id: 1,
 			timestamp: new Date().toISOString(),
 			amount: 10,
 			username: "test-user",
-			tree_id: "00008100:002d8af7",
+			tree_id: "00008100:00284266",
 		},
 		{
 			id: 2,
 			timestamp: new Date().toISOString(),
 			amount: 20,
 			username: "test-user",
-			tree_id: "00008100:002d8af7",
+			tree_id: "00008100:00284266",
 		},
 		{
 			id: 3,
 			timestamp: new Date().toISOString(),
-			amount: 15,
+			amount: 30,
 			username: "test-user",
-			tree_id: "00008100:002d8af7",
+			tree_id: "00008100:00284266",
 		},
 		{
 			id: 4,
 			timestamp: new Date().toISOString(),
-			amount: 5,
+			amount: 40,
 			username: "test-user",
-			tree_id: "00008100:002d8af7",
+			tree_id: "00008100:00284266",
 		},
 	];
 
 	const treeData: TreeCoreData = {
-		id: "00008100:002d8af7",
-		lat: "13.54072",
-		lng: "52.57403",
-		art_dtsch: "Kulturapfel",
-		art_bot: "Malus domestica",
-		gattung_deutsch: "APFEL",
-		gattung: "MALUS",
-		pflanzjahr: 2021,
-		standalter: "2",
-		baumhoehe: "3",
-		bezirk: "Lichtenberg",
+		id: "00008100:00284266",
+		lat: "13.57305",
+		lng: "52.44338",
+		art_dtsch: "Spitz-Ahorn 'Summershade'",
+		art_bot: "Acer platanoides 'Summershade'",
+		gattung_deutsch: "AHORN",
+		gattung: "ACER",
+		pflanzjahr: 2019,
+		standalter: "4",
+		baumhoehe: "7",
+		bezirk: "Treptow-Köpenick",
 		eigentuemer: "Land Berlin",
-		radolan_sum: 153,
+		radolan_sum: 191,
 		geom: {
-			coordinates: [13.54072, 52.57403],
+			coordinates: [13.57305, 52.44338],
 		},
 		radolan_days: [
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 3, 1, 0, 2, 6, 9, 0, 1, 0, 4,
-			4, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 4, 3, 1, 2, 7, 7, 1, 4, 2, 6,
+			5, 2, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-			0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 5, 0, 0, 0, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 3, 1, 4, 0, 0, 0, 0, 0, 1, 0, 0, 2, 7, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 12, 2, 0, 0, 0, 0, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 2, 17, 8, 1, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 2, 2, 2, 2, 1, 0, 0, 5, 2, 5,
-			8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-			0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 2, 0, 0, 0, 0,
-			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 7,
-			7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 3, 0, 0, 0, 0, 0, 6, 5,
+			11, 12, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 7, 2, 0, 0,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
+			4, 7, 9, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 13, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 			0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
@@ -94,26 +94,28 @@ test("should calculate correct water needs for baby tree", () => {
 		treeData,
 		new Date("2024-01-01T00:00+00:00"),
 	);
-	expect(treeAge).toBe(3);
-	expect(treeAgeClassification).toBe(TreeAgeClassification.BABY);
+	expect(treeAge).toBe(5);
+	expect(treeAgeClassification).toBe(TreeAgeClassification.JUNIOR);
 
 	const {
 		rainSum,
 		rainPercentage,
 		wateringSum,
 		wateringPercentage,
-		otherWateringPercentage,
 		referenceWaterAmount,
 		shouldBeWatered,
 		waterParts,
+		stillMissingWater,
 	} = useTreeWaterNeedsData(treeData, waterings, treeAgeClassification);
 
-	expect(rainSum).toBe(15.3);
-	expect(wateringSum).toBe(50);
-	expect(referenceWaterAmount).toBe(100);
-	expect(rainPercentage).toBe(0.153);
+	expect(rainSum).toBe(21.1);
+	expect(wateringSum).toBe(100);
+	expect(referenceWaterAmount).toBe(200);
+	expect(rainPercentage).toBe(0.10550000000000001);
 	expect(wateringPercentage).toBe(0.5);
-	expect(otherWateringPercentage).toBe(0.347);
-	expect(shouldBeWatered).toBe(false);
-	expect(waterParts.map((p) => p.progress)).toEqual([0.153, 0.5, 0.347]);
+	expect(shouldBeWatered).toBe(true);
+	expect(stillMissingWater).toBe(79);
+	expect(waterParts.map((p) => p.progress)).toEqual([
+		0.10550000000000001, 0.5, 0,
+	]);
 });
