+++ tests/unit/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+// eslint-disable-next-line @technologiestiftung/no-default-export
+export default defineConfig({
+	test: {
+		include: ["./tests/unit/**/*.test.ts"],
+	},
+});
